Update and fix ConsistentPolicy.
There was a problem with the initializer with some weird logic where it
would completely ignore whatever nodes were passed to it.  It would use
the values for the times when it shouldn't.

In addition, the ConsistentHash was unnecessarily performing two
lookups.  This goes against the general idea of how the consistent hash
is meant to work.

Finally, since the Policy files updated so far have not provided an
implementation for the printAdditionnalStats()[sic] method, a
placeholder implementation was put in Policy so that subclasses can
ignore the method.