Don't assign Node IDs or Networks in constructors.
Update the examples to reference the 0.9.1 version of the framework.

Because setting the Network in the constructor doesn't really provide
any benefits, that parameter was removed.  In addition, the IDs are now
dynamically generated.  When they are needed in the code, they can be
retrieved.  This actually makes for cleaner code where it is more
obvious how connections are being made between various elements.